Refactor primitives.
