+++ codec/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "polkadot-runtime-codec"
+name = "polkadot-codec"
 description = "Serialization and deserialization codec for runtime values"
 version = "0.1.0"
 authors = ["Parity Technologies <admin@parity.io>"]
