+++ codec/src/endiansensitive.rs
