+++ codec/src/joiner.rs
