+++ native-runtime/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 authors = ["Parity Technologies <admin@parity.io>"]
 
 [dependencies]
-polkadot-runtime-codec = { path = "../runtime-codec", version = "0.1" }
+polkadot-codec = { path = "../codec", version = "0.1" }
 polkadot-runtime-std = { path = "../runtime-std", version = "0.1" }
 polkadot-primitives = { path = "../primitives", version = "0.1" }
 rustc-hex = "1.0"
@@ -13,4 +13,4 @@ log = { version = "0.3", optional = true }
 
 [features]
 default = ["std"]
-std = ["polkadot-runtime-codec/std", "polkadot-runtime-std/std", "polkadot-primitives/std", "log"]
+std = ["polkadot-codec/std", "polkadot-runtime-std/std", "polkadot-primitives/std", "log"]
