+++ runtime-std/Cargo.toml
@@ -13,7 +13,7 @@ pwasm-libc = { path = "../wasm-runtime/pwasm-libc", version = "0.1" }
 environmental = { path = "../environmental", version = "0.1", optional = true }
 polkadot-state-machine = { path = "../state-machine", version = "0.1", optional = true }
 polkadot-primitives = { path = "../primitives", version = "0.1", default_features = false }
-polkadot-runtime-codec = { path = "../runtime-codec", version = "0.1", default_features = false }
+polkadot-codec = { path = "../codec", version = "0.1", default_features = false }
 triehash = { version = "0.1", optional = true }
 ed25519 = { path = "../ed25519", version = "0.1", optional = true }
 
@@ -24,7 +24,7 @@ std = [
 	"polkadot-state-machine",
 	"triehash",
 	"polkadot-primitives/std",
-	"polkadot-runtime-codec/std",
+	"polkadot-codec/std",
 	"ed25519",
 ]
 nightly = []
