+++ wasm-runtime/polkadot/src/runtime/genesismap.rs
@@ -20,8 +20,7 @@ use std::collections::HashMap;
 use runtime_std::twox_128;
 use codec::{KeyedVec, Joiner};
 use support::Hashable;
-use primitives::relay::{Number as BlockNumber, Block};
-use primitives::AccountId;
+use primitives::relay::{Number as BlockNumber, Block, AccountId};
 use runtime::staking::Balance;
 
 /// Configuration of a general Polkadot genesis block.
