+++ wasm-runtime/polkadot/src/runtime/staking.rs
@@ -21,7 +21,7 @@ use runtime_std::cell::RefCell;
 use runtime_std::print;
 use codec::KeyedVec;
 use support::{storage, StorageVec};
-use primitives::{BlockNumber, AccountId};
+use primitives::relay::{BlockNumber, AccountId};
 use runtime::{system, session, governance};
 
 /// The balance of an account.
@@ -216,7 +216,7 @@ mod tests {
 	use runtime_std::{with_externalities, twox_128, TestExternalities};
 	use codec::{KeyedVec, Joiner};
 	use support::{one, two, with_env};
-	use primitives::AccountId;
+	use primitives::relay::AccountId;
 	use runtime::{staking, session};
 
 	#[test]
