+++ wasm-runtime/polkadot/src/support/environment.rs
@@ -21,8 +21,7 @@ use runtime_std::mem;
 use runtime_std::cell::RefCell;
 use runtime_std::rc::Rc;
 
-use primitives::relay::{BlockNumber, Digest};
-use primitives::Hash;
+use primitives::relay::{BlockNumber, Digest, Hash};
 
 #[derive(Default)]
 /// The information that can be accessed globally.
